// source/frontend/src/models/api/generated/ApiGen_CodeTypes_ConsultationOutcomeTypes.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
export enum ApiGen_CodeTypes_ConsultationOutcomeTypes {
  APPRDENIED = 'APPRDENIED',
  APPRGRANTED = 'APPRGRANTED',
  CONSCOMPLTD = 'CONSCOMPLTD',
  CONSDISCONT = 'CONSDISCONT',
  INPROGRESS = 'INPROGRESS',
}